s5proxy is aware of the active vs. desired SSH proxy connection, + other supporting changes.
While processing the "on" command-line option, an existing SSH connection is verified to be to the correct host. AN incorrect SSH connection is killed, and the correct connection is started. This helps when the $ssh_host variable has changed between script invocations.

Minor comment cleanup, at the top of the script where variables are set.

Defined $ssh_proxy_command_stub, to contain all of $ssh_proxy_command excluding the SSH hostname. This lets the script find the ssh process, without knowing the hostname which was connected to. This helps kill a previous SSH proxy connection, when the $ssh_host variable has changed between script invocations.

Defined proxy_settings_differ() which returns true if the settings in OSX Network Setup are different from those in the script. This allows the script to only change Network Setup only when necessary.

Defined proxy_on(), which returns true if the proxy is already enabled. This skips an unnecessary modification of Network Setup, and prompting for a sudo password. This also helps when the $ssh_host variable has changed between script invocations.

Defined get_ssh_proxy_pid(), which sets a global $ssh_proxy_pid variable. This PID-finding code was broken out into it's own function because it's now used in more than one place in the script.

Defined get_ssh_host(), which sets the global $found_ssh_host variable. This is used to determine whether a current SSH proxy connection is to the same host which is set in $ssh_host. This helps the script know whether it needs to kill an old SSH connection and establish the correct one.

kill_ssh_proxy() now uses get_ssh_proxy_pid, $ssh_proxy_pid, and $ssh_proxy_command_stub.

enable_proxy() now clarifies that you will be prompted for your sudo password. This makes it a bit less confusing when you are prompted in succession for an SSH password, then a sudo password.

disable_proxy() now clarifies that it will leave the rest of the settings in Network Preferences intact, while disabling the proxy.

proxy_status() uses get_ssh_proxy_host and $found_ssh_host to show which host is actually connected to.

proxy_status() clarifies which interface we are talking about, since there are Network Setup settings for multiple interfaces.

While processing the "killproxy" command-line option, clarify that network traffic will be interrupted when the proxy is enabled, but no SSH connection is running.

While processing the "setup" command-line option, only adjust Network Setup preferences, if they differ from the desired settings.