expandtemplate/expandtemplate uses double brackets [[___]] now, to represent variables, and processes multiple templates on the command-line.
Expandtemplate uses double-bracketed strings, instead of single-bracketed strings, to represent variables. This was changed to allow use of single-bracketed strings within templates, E.G. markup.
expandtemplate/sample (The sample template) was updated to use double-bracketed variable names.
Expandtemplate now allows multiple templates to be specified on the command-line - if the same variable is mentioned in multiple template, that variable is only prompted for once.