git-trix/set_git_author_from_parent_shell defaults email_domain, works in OSX, and only sets variables when another non-root user is found.
The email_domain variable defaults to the last two components of the system's domain where this function is running.

Git_* variables are only set when a different user which is not root is found.

The process tree stops being walked, if the PID is < 2, as we are out of processes and not going to find a different user.

The ps command now uses "-o user= -o ppid=" to be compatible with both Linux and OSX while not displaying a header, vs. the Linux h (no headers) ps option which is unsupported in OSX.

In OSX, use the dscl command to get the full name of the user, as there is no getent command.

Added some comments inside the function, to explain it's purpose, and some of the logic.